icons: break out get_icon_for_gicon and use ICON_LOOKUP_FORCE_SIZE
The force-size flag is new for gtk 2.14, and means we always get a
pixbuf the size we ask for!

We use get_icon_for_name, which now takes a whole row of names, that
we try after each other; if any icon is found, it is stored in the
cache under the first name-- to speed up identical lookups.