wscript: Optional dependency gnome.ui
To connect to the desktop session for Gnome 2.22 we have to use
gnome.ui (aka libgnomeui) which is deprecated technology. It is only
used conditionally and links to it dynamically, which is very easy
with python. Currently that is the only way to do session connection
right now.

This method with GnomeClient from gnome.ui should work on Gnome 2.26
as well, provided gnome.ui is installed.