browser: Do not reset after activation (among other)
A regression was that we reset after text entry changes to ""
regardless of the cause; we regress to clear the search after
launching an item (it should stay as default for next action).

At the same time, we have to have a solution when the field is cleared
in other ways; we cover the Ctrl-backspace case by hardcoding it to
clear the field. This is unfortunate, since we override normal
Ctrl-Backsp behavior = Erase word, and we hardcode it as well.