data: check if search results are valid before they are used
Since matches are mostly valid, we apply a generator filter
valid_check on the search results, so that only valid search results
are returned.

The iter is only used on the items actually used on the model side, so
that we only do the check on the first item, or only the items
displayed in the hit list.