browser: Use two-line list layout with thumbnails
Make sure to escape strings for markup in text label in list

Simply use get_thumbnail and get_pixbuf in result list, so that all
items get their real icon.

The actions benefit from using very simple icons in the result list,
while some objects, such as music albums and pictures, benefit greatly
from thumbnail pictures in the list.

Composed icons however, are not suitable, and we will device a
solution so that we don't use them in the list.