icons: Remove all imports in functions
The icon module needs the common modules gtk and gio that are used
all over, it does not save anything to import them later.

Items that are needed very often should be imported into the module
namespace, for the rest just using the module is good enough (when
only being called a few times). For example, the ThemedIcon etc
symbols are good to import since they are used often.