objects: OpenWith, better description and icon
icon should use application icon when small, so composed icon gets the
emblem as fallback.

We don't return a description as that clutters and is just a
duplicate of the title

We don't need to write out that an application is  "default", the user
will understand, the action is named "Open" and is ranked first.