data: Search asynchronously by default
Search with a short timeout delay in the mainloop. Create a new
infrastructure to keep track of the current search in each pane and
terminate the ongoing search when a new one is posted if any.

Apply a heuristic in DataController.search related to the key length;
if len(key) is 0, we work interactively and browsing does not cost
anything, so timeout is 0. Then we have a long timeout for len 1, and
decreasing timeouts for longer searches (as the number of matches
shinks, it's much faster to search).