plugin.rhythmbox_support: Keep string mapping and reuse strings
Make sure we reuse strings; we will have lots of duplicate artist and
album names in the info dictionaries for each song; we keep a map in
the content handler to make sure equal strings use the same instances.

We also reuse the wanted-keys collection to collate element names as well

This is estimated to save 1-3 megabytes of heap space for a large
library (3000 songs).