plugin_support: Handle builtin settings
We have to copy all settings description dictionaries, so that they
are all unique.

If it is a core setting, we write it back on plugin initialization if
it was not set; thus, plugins can set default values for core
settings.

Check the section in the value-changed signal callback; we can only
update our value if it belongs to our plugin.