plugin.favorites: Fix Registering and Finding the FavoritesSource
We keep a list of weak references to all FavoritesSource ever created;
the only referenced one of these is the FavoritesSource we want to
work with.

At any time, we should have at least one source, and we only have more
just before replacing created with unpickled, or just when creating
an instance to get its icon.