objects, helplib: Move Mixins  to helplib
We collect helper constructs in helplib, so we move the
PicklingHelperMixin and the FilesystemWatchMixin there.

The Mixins are still imported into objects.py; they are needed there,
but plugins importing them there will not have cache breaks either.

We also rename to _nonpersistent_token to NonpersistentToken; we fix
this ugly classname in c13; then it has never been part of any release
(apart from stillborn c12).