Introduce qfurl: kupfer URIs and URI resolution
A qfurl is a URI to locate unique objects in kupfer's catalog

The qfurl is built up as follows:
``qpfer://mother/qfid#module_and_type_hint``

The mother part is a mother source identifier and is optional.
The module_and_type_hint is optional.

A short url looks like the following:
``qpfer:identifier``

This class provides methods to get the qfurl for an object,
and resolve the object in a catalog.