relevance: Simplify _findBestMatch loop
Apparently, checking if we got the whole query in the end of the
string checks three conditions of which the latter two are always
true. Doing less work here, and in addtion exiting directly when the
query doesn't fit anymore, saves some cycles.

Save len(query) in local variable, since we use this measure three
times every turn of the loop, we can save some cycles by precomputing
it.