relevance: Rewrite _findBestMatch inner loop
The inner loop should loop over the query, not the string since the
query is always shorter than the string (or equal).

The inner loop will loop over the query characters, trying to find
them between the last found query char and the known end of the range.

This speeds up string ranking to wicked fast levels!