browser: Use object's real name in the results list
While we show the actual alias that was matched in the match view (for
example gedit), we show the canonical/true name of the object in the
results list (Text Editor).

Now the results list always shows the same item under the same name,
offers insight that the matched alias is only an alias, and is less
ugly in case the matched alias is a folded name.