data: Check if an Action is valid when displayed
We postpone valid checking for actions (just like for leaves) until it
is displayed; Only the first item is normally checked. This avoids any
action checking that is not immediately needed (to display the
first/default action). When browsed or searched, the actions are
immediately checked passing through the iterator.

We cache the result for a particular item.