objects: Resolve real path of FileLeaves lazily
If we call os.path.realpath at construction time for FileLeaf (to
resolve real paths from symlinks), we incur lots of I/O upon browsing
directory sources, and effect that slows down even if the directory
listing is relatively small.

Instead of letting FileLeaf.object be a resolved path, it is the path
that was used.

We add FileLeaf.canonical_path() to get the real path. We use this for
the description, and we use os.path.samefile(f1, f2) for FileLeaf's
__eq__ method.