plugins: Load a Fake Plugin for plugins that raise on import
Plugins should raise ImportError on import, and Kupfer's plugin loader
will register it and record the error. ImportErrors are special-cased
and any other errors raised on plugin import (such as SyntaxError) are
printed with a traceback instead.

After we catch the ImportError we use a minor hack to still try to get
as much information as possible out of the plugin; we ask the import
loader for its code object and evaluate it; thus we can grab the
kupfer information variables in the plugins, if they are defined
before the exception is raised.