plugin.virtualbox: Import without catching ImportError
Generally, kupfer's Manual.rst says this:

    The plugin should not do **anything at all** upon module load,
    except loading its required modules. Load modules without
    try/except; ImportErrors will be caught by the plugin loader and
    the plugin disabled

This makes the plugin much simpler to implement. However the drawback
until now has been that the plugin would not even show up in the
plugin list. Now with some experimental changes I inject a FakePlugin
object if we encounter an ImportError when loading a plugin.

The FakePlugin can read in the plugin name and description to be shown
in the plugin info, and the error will be displayed prominently.

This means that plugins can now use bare imports, without catching
ImportError, but they have to do this after definining the special
kupfer plugin attributes, so that those lines can be executed before
the exception is raised.