plugin.opera: Adjust error handling
Use try..except handling of the bookmarks file to catch any cases for
example nonexistant directory or permissions problems.

Explicitly convert the name to a unicode string; the opera bookmarks
file defines an encoding (oh horror to parse and adapt to that),
instead we assume it is UTF-8 which is the default for Opera.

Also adjust strings.