plugin.locate: Respond quicker with first results
Locate does not streamingly output to stdout if it is connected to a
PIPE, so we launch two locate processes -- one with limit 10 and one
that goes to the maximum limit.

This way, we can show the first results relatively quickly, instead of
waiting for locate to return the maximum number of results.