plugin.notes: Format text when creating new notes
Try to format text to make a sensible title. If the first line is a
good title, pass the text on as it is, the first line will become the
note's title and the following text the content.

If the first line is too long, we break it off and take the first
words as a title while keeping the full text as content to not break
any sentences between title and content.

This works well for latin languages. Doing smart things with other
languages can be implemented later.