main: Call locale.bind_textdomain_codeset
For gtk.Builder, we need to call the C library gettext functions
As well as set the codeset to avoid locale-dependent translation
of the message catalog.

This means that all gettext message catalogs for kupfer need to be
UTF-8, since this is hardcoded at this point. Only the gtk.Builder
interface parts depend on this.