plugin.epiphany: Parse Epiphany bookmarks with ElementTree
Remove lots of old code that we never used for Favicon parsing and
History parsing (we can retrieve it in the history later if we need
it).

Parse Epiphany's RDF bookmarks file with ElementTree, this makes the
code smaller and more parts of kupfer use the same XML parsing
package.