plugin.pidgin: Listen to Sign Off signal and rework signals
Listen to SigningOff signal to remove contacts as whole accounts sign
off. This signal is also called when Pidgin quits normally.

Implement this with _remove_buddies_not_connected, which removes all
buddies whole account is not connected right now.

We also listen to Buddy Connected and Status Changed signals to get all
buddy updates. Single buddy updates are added to a queue and scheduled
a short time in the future. This has two main benefits:

* Many buddy updates per connection is more effective, and we only
  process each buddy at most once.
* Buddy update happens outside the D-Bus signal notification, so we
  are safer during extreme events (signals during Pidgin teardown)

Insert some debug printouts so that we see where we do (too)
frequent updates or double updates.

Use the helplib's convenience function for connecting to dbus signals.