plugin.google_translate: Change action name
This is just a *suggestion* for action and object names that align
better with Kupfer's current customs.

An action that requires an indirect "third" object should mostly have
a name ending with "..."

If we name the language objects simply "English" we can read the
action like a sentence:

	(Block of text) -> Translate To... -> English

This mimics the common SVO sentence structure of many languages.
Does it fit well with Polish?