plugin.pidgin: Tell pidgin to present conversation in OpenChat
We try to focus the new conversation in the OpenChat action by telling
Pidgin to present it.

There are some problems that are not easy to work around in Kupfer:

Tabs do not switch automatically, this is a bug in Pidgin. Also,
window focusing does not work perfectly with already open chat windows
due to "focus stealing prevention".