contrib: Add icon_names plugin for browsing icons
This is just a very quick hack to be able to search the standard
icons to find icons and see what they look like.

Plugins in contrib/ are things that someone might find useful or
interesting to study, but are not appropriate to include in Kupfer.

This plugin might be useful in developing Kupfer, but it's not
generally useful. It should also be made to browse more than just the
standard icons.