plugin.google_translate: Use and catch specific exceptions
In Python 2.5, GeneratorExit inherits from Exception, so we get errors
like these:

    Exception exceptions.RuntimeError: 'generator ignored GeneratorExit'
    in <generator object at 0x10fda418> ignored

GeneratorExit is raised inside the generator when it is deleted before
it finishes, so that our finally: clause runs.

We fix this to only catch HTTPException and ValueError (the latter
which we use internally).

This problem does not occur in Python 2.6, where GeneratorExit only
inherits BaseException.