plugin.locate: Remove is_dynamic
With a change in Source, we can now yield items as we receive them,
while still being cached. This allows the locate action to display the
first 10 hits first, then the rest later (if the user scrolls down the
list or tries to filter the list, all items must load).