helplib: Add reverse_action to helplib
reverse_action is a convenience function to reverse a three-part
action.

This type of "reversal" is not done automatically in Kupfer and I
don't think it can be done either; some three-part actions should
never be reversed (like  Document -> Move To.. -> Folder becomes very
confusing in the form Folder -> Move Something To(?).. -> Document).