Start kupfer with kupfer.py
We make a script kupfer.py that will always be used to start kupfer.
This script is placed inside $PYTHONDIR, and we save a reference to
PYTHONDIR in the kupfer-activate.sh script, so that we can find where
the kupfer Python package is installed.

With this change, it is possible to install kupfer's python package
anywhere, for example inside $DATADIR/kupfer like the Debian package
for Kupfer does. That way, we don't have any issues with collisions
with other system packages, or Python's sys.path at all.