wscript: Install into DATADIR unless --pythondir-install
Change kupfer to default to being installed with all its Python
package into DATADIR.

The Python package is installed as arch-independent files into
DATADIR, which has support in the FHS, and is widely practiced for
Python packages in Debian. Since Kupfer does not include optimized
.pyo by default, we avoid the issues with these (not
arch-independent).

The configure option --pythondir-install is added to allow installing
into the old location, as a system-wide Python module.