plugin.text: Allow open URLs in Text
Add action OpenTextUrl to open web addresses in Text. At the same time
we "reduce" the URL matching in free-text to produce UrlLeaves only
for "real" URLs (with http:// etc) while fuzzy URLs will just be
treated as text (now with OpenTextUrl action).

This makes for less surprises if the user *doesn't* intend to type in
an URL.