plugin.gajim: Listen to D-Bus signals
Try to make the Gajim plugin only update on changes, so that we don't
have to call over d-bus every time we do a new search. This can save a
little time when Kupfer is interactive; in extreme cases, the user may
notice delay in each search with the old behavior.