plugin.calculator: Try to fix paranthesis deficit in expressions
When given an expression like "log(2" we try to be smart and add the
missing brackets at the end of the expression.

I can't think of this breaking anything at all, except paranteses
inside string literals, but we don't support any way to use string
literals at all. (You *can* do math with string literals and this
example will fail with new code: '"(".count("(")' -- should return 1,
but is really bogus.)