data: Check all sources; reload even dynamic sources on load
In SourceController.cache_toplevel_sources, cache even dynamic sources
-- this gives Kupfer a chance to catch exceptions early and disable
the relevant plugins.

This fixes a change in behavior from c17 to c18, now the Special Items
(Trash functionality) will be disabled properly again if it raises on
first refresh.

The real fix for users is to install gvfs.

With thanks to Jan for providing info to help fixing this bug.