browser: Fix lazy loading of result list
Upon every step down we check if we have enough items in the result
list to display the desired scroll amount. We should check this
exactly, else we load in 10 new result for each down key pressed.
Loading in too many results is a performance issue.

Now we only load in if current row index plus requested scroll amount
is more than the number of loaded items.