contrib: Allow browsing all icons in icon_names
icon_names is a handly tool but not a real Kupfer plugin (not for
enduser use). now it can browse all icon names in the gtk icon theme.

Note: not any icon name found in the current theme can be used for
Kupfer, we must always try to use icon standard names or gtk stock
icon names when possible.