plugin.trash: Hide toplevel trash source
It is not completely optimal, but we reuse the InvisibleSourceLeaf
trick for the trash plugin. This makes the trash item behave just like
it did before (as an object in the catalog, not a catalog itself).

I personally see no use of having the *contents* of the trash in the
toplevel catalog, so no reason to make it possible.