browser: Don't show and unshow the third pane
If we use WIN_POS_CENTER to only influence the *very first* placement
of the window, then reset it to WIN_POS_NONE, then the window won't be
recentering like before. Thus we don't need to do the hide/reshow
dance with the third ("indirect object") pane for Kupfer (when it's
visible).

This means Kupfer's window finally behaves like we always wanted --
center on first appearance, and simply stay put (wherever the user
moved it) after that.

This may potentially fix Launchpad bug
https://bugs.launchpad.net/kupfer/+bug/494237

With luck it may even fix another bug,
https://bugs.launchpad.net/kupfer/+bug/417278