objects: Treat AppLeaves created from files better
If we create an AppLeaf from a file (init_path=..) we have to take
care about the application's .get_id() method. File-based AppLeaves
now use the file path as ID.

This consolidates the application id methods to the new function
kupfer.launch.application_id() that takes a gio.AppInfo object as
parameter (and uses a trick: we "decorate" the objects we parse with
an extra attribute with the file's path).

We also include the path in the description so that you can
differentiate if you have many similar/equal .desktop files lying
around.

This solves an issue reported by Chmouel (I hope).