browser: Add hints to qpfer URL lookups for selection
Add type hints to the URL lookups for selected file etc. This
dramatically narrows the number of sources we need to search for the
items.

The issue is mostly not lookup speed (which is quick), but that we
might hit a source that must be rescanned to be searched; and that may
slow the lookup by a comparatively big factor.