compose: Add keybinding Control+Return to compose Object + Action
Composition allow the user to make one Kupfer object out of an
(Object, Action) pair or an (Object, Action, Indirect Object) tuple.

If we allow the user to create objects out of these combinations, we
can later add "higher orders" of Kupfer usage, for example -- run
action at (later) time, add alias or trigger for this action, etc.

At the moment, we only allow to run the composed object, and thus the
feature does not have much use (yet).