plugin.windows: Remove gtk event time assert
This assert has helped to make sure we correctly activate window
actions inside a gtk event and have the right interactive behavior.
However, this gtk event time definedness is now (sometimes) violated
if we use the Run After Delay... action.

So we remove this restriction, even though it deserves to be checked
also in the future.