plugin.core: Internalize Selection and Text
We make the very basic plugins for current text selection and for
Free-text Queries built-in parts of the core plugin package.

These plugins are too light to stand on their own, and in normal
Kupfer use it hardly makes sense to be able to remove the text.py
plugin.

The core plugin *package* makes it rational to internalize more
functionality to be "builtin" to Kupfer (by virtue of being part of
the hidden, enabled-by-default core plugin).

This allows us to remove some noise on the string side; these
internalized plugins don't need localized plugin descriptions etc.