browser: Hide third pane in .put_away(), show on activate
This addresses and fixes bug https://bugs.launchpad.net/bugs/498542
that was reported by Chmouel.

To suit most window managers, we try to hide the third pane (if
visible) on put_away and show it again later. Thus the window is
"always the same size" and the window manager will do the right thing
and let it stay in the same place.