Add a triggers plugin.
With the remove trigger action, the trigger handling is almost
complete. This plugin is highly experimental but already works
surprisingly well;  you notice here that this plugin has no preference
pane but only uses Kupfer's "normal" interface -- this is since it is
so convenient and easy to add to Kupfer using this method if you are
used to it.