core: Add debug tools 'Forget' and 'ComposedSource'
Add an action to "Forget" about an object: erase its leared score in
the 'learn' module.

We also add ComposedSource which makes Composed Command objects
transparent, so that we can access individual objects inside them.
This is just an experimental feature.

Also, we make sure the whole subplugin 'core.debug' is only loaded
when Kupfer is in debug mode; so in normal mode, the module is never
even imported.