puid: Resolve actions by reference
Resolve actions by reference: it's quite simple, the catalog is very
small. Support resolving builtin actions if a leaf is given, else we
look in the action decorator "library", which also will just look at
currently loaded plugins.

This is so that less objects are serialized by general. This way, we
never need to serialize actions.