objects: Fix repr_key for OpenWith and SourceLeaf
And OpenWith does not need to be pickleable anymore.

If the OpenWith (default) action is now resolved by reference, it will
always find the *current default* open with action. Which can be
expected in a way. If the user wants to save an absolutely specific
trigger, you should use the Open With.. (any) action.