browser: Never show an empty match. Keep latest result.
We never show an empty match for a search query, but keep the latest
result.

The Hypothesis is that the user does not gain from "No action for
'..'" or "No matches for '..' inside '..'" -- if you type a good query
but with a wrong letter at the end, it can be frustrating.

With this change, we simply "freeze" when no matches are found for a
query; we simply refuse updating the interface and keep the latest
match(es) displayed.