grouping: Use JID as email address in Pidgin if valid
If the JID is a valid email address we register it as such for Pidgin
contacts. This is twofold: We allow e-mail actions to work on pidgin
contacts, which is really flexible. We also use the address for
matching/grouping with other contacts.

Grouping API updated to ignore values of None when grouping leaves.