Add 'Close' action on AppLeaves to close all windows
We use window matching to try to find all windows of the app and call
close on them.

There is no Application API on Linux, which is sad. We can't call this
action Quit; we have no way to quit an application in general. Many
applications will quit if we close all windows, but far from all.
Calling this action "Close" will catch the concept.