google_translate: Make possible to save triggers with translate
We mark the _Language class as serilizable; Kupfer will then save a
representation of the Language object right in the trigger (since it
can't find the language in the catalog). The only important part saved
is the _Language.object attribute.

If the plugin changes incompatibly in relation to _Language, you can
increase the 'serilizable' number: if it doesn't match the saved copy,
the serialized trigger won't be active anymore.