obj.base: For Actions, use Action's icon name as last fallback
Unify KupferObject.get_icon and .get_pixbuf to do the same thing.
get_pixbuf is kept as is, since it is used more frequently and avoids
disk accesses (by not asking if icons are good/files exist).

We make sure that for actions that don't find their icons, we use the
default action icon. Other objects unchanged.