core.sources: Take source name into account for Cache file name
We change the cache file name (again) so that the name of the source
is part of the generation. Then sources are automatically changed when
we switch locale.

We also bump the cache version number to 4!

This means after this change, **Kupfer will erase all** old cache
files with version number 3 in their name, to clean up.