Alternative ranking for Actions
This adjusts the rank for actions, *without query* (no search text),
to be more rigid. Basically, we honor all rank_adjust actions, so
default actions are almost always on top, and negatively adjusted
actions are always at the bottom.

This is the only way we can get sane and predictable results right
now, what is needed for the future is:

 * Object-specific action memory
 * Action configuration: which actions to use, which rank to give them

With a search query in the action pane, it behaves exactly as before,
with the old ranking mechanism.