fileactions: Allow creating archives from a collection of files
We port the archive actions in the File Actions plugin to the multiple
dispatch protocol. This means that with multiple selected files, the
actions now allow to create one archive with all the files.

This is what the comma trick + multiple dispatch is made for, now it's
possible to select some files in nautilus, invoke kupfer's magic
keybinding (or open kupfer, type Ctrl+G) and go
Selected Files -> Create Archive.