rhythmbox: Adapt "Play" action to multiple dispatch
Adapt the PlayTracks action to multiple selection ("comma trick") to
play all songs in order (play the first, enqueue the rest).

The amazing thing is that now that it is possible, using the play
action on all my 3000 songs in the library does work, only takes 3
seconds and adds all songs to the play queue.