ui: Delay text mode change to postpone window resize at activate
We attack the problem of spontaneous blanking (only when kupfer's
window resizes) by putting the reset/end text mode after activate in
the Interface in an idle callback. It seems that this way we do not
blank(!).

Activating directly from text mode always caused window blanking with
a compositing window manager.