apt_tools: Use apt-get install --yes  as default install method
It is unfortunate, but I think we have to use the --yes flag on the
apt plugin for this to work properly. We want to use gksu so that the
user can verify the commandline being run before authorizing it.

With gksu the terminal is somehow then locked, and apt-get can't
accept user input; we have no choice but to either use sudo or use the
--yes flag to apt-get.