core: Guard plugin loading and guard plugin intialization better
We guard plugin loading so that plugins that raise exceptions when its
attributes (sources, actions etc) are instantiated are disabled safely
and don't cause Kupfer to stop functioning.

We also guard source initialization and remove content decorators from
the same plugin if we raise an exception.

This solves the current evilplugin testcase, so that this debug plugin
can be installed without breaking kupfer.