Localize helper script kupfer-exec
Use bash localization to localize the single message in kupfer-exec
that must be translated.

Not finding a good way to let intltool extract the strings inside
kupfer-exec, we simply define the same strings in kupfer/version.py to
put them in the catalog.

Note: you can extract manually with::

    bash --dump-po-strings bin/kupfer-exec

Also notice that the order of definition for the TEXTDOMAIN and
TEXTDOMAINDIR variables is important.