Assign icons to saved kupfer command files
Assign icons by taking the Kupfer icon pixbuf and writing as a
thumbnail for the file. Then we set the metadata property
metadata::custom-icon to point at the thumbnail file, which makes sure
the icon stays with the file as it moves around.

To make this work in the long run, we also always check on an executed
command file if we can refresh its icon.