Put files from commandline in kupfer directly as files
When invoked with a list of files, kupfer FILE [ ... ], we put them
directly in kupfer as files on the stack. Thus we support putting
multiple files in Kupfer at the same time.

The only way to put text queries in kupfer from the command line is to
pipe text into the ``kupfer`` command.

This is implemented with D-Bus method changes, where PutText now only
takes a string, and PutFiles takes an array of file uris.