"Convert" OperationErrors to user error notifications in commandexec.py
When executing a command and encountering OperationError, we show an
error already in commanexec.py (so we get an error regardless of from
where the command was executed, including triggers).

Error "conversion" means that we consume the error (if we succeed in
showing an error dialog for it), but raise a more general action
execution exception.