Add a filter to clean out comments in *.po translation files
We enable a filter for *.po files, that can strip out
information in the *.po that is not needed in to save in the git
repository.

To enable this for a language you must add a line like this to the
.gitattributes file, replacing 'sv' with the actual language:

    /po/sv.po               filter=cleanpo

While the .gitattributes file marks the *.po files to use the
"cleanpo" filter, you have to actually define the "cleanpo" filter in
your repository to make it work. If you want to commit translations
without the filename and line number comments saved to the repository,
you need to add the following to your .git/config configuration file:

    [filter "cleanpo"]
        clean = grep -v -E "^#: "

This filter will remove the line number and filename information from
the po files, so the stored diffs will be smaller and easier to read.