twitter: Friend: contact group takes precedence over timeline as content
Since the Friend class is a ContactLeaf, and the friend source is a
Grouping Source, it may be in the situation where it is grouped
together with other contacts, and the twitter Friend is the "main"
contant.

In this case, we must delegate to the superclass' implementation of
has_content and content_source. In this case, you can access the
timeline by first entering the Twitter contact, then its "real" entry
inside itself. It doesn't sound too logical, but I think it will work.