If we encounter a module with 'None' code, return early
If we have a stale .pyc (or .pyo?) file from the wrong Python file,
and no present .py file for a plugin, then loader.get_code() for that
module will return None. If we get a None code object we have to skip
that module.

This fixes a bug reported by Chmouel Boudjnah

https://bugs.launchpad.net/bugs/544289