google_translate: Use ajax APIs
With thanks to hetdegon. Use the generic APIs supplied by google for
getting a JSON response. This format is expected to be stable and thus
much safer for Kupfer to use.

Dictionary mode is unavailable when using the API right now.

Work in Progress: Need HTTP referrer and API Key to use the API

http://code.google.com/apis/ajaxlanguage/documentation/reference.html