commands: Add Action Pass to Command
Pass to Command: Run a command (executable with optional arguments)
and supply the text or file object as additional parameters.

Comma trick is supported on the object side to pass multiple
parameters (texts or files).

The increased power and open-endedness of adding capabilities like
this to Kupfer leaves lots of questions and choices and throw away
some simplicity. Which actions should save command output and which
should not? How should actions react on errors etc?