Port all plugins to use utils.spawn_async
Even the simple parsing of commandlines is worse than no parsing at
all. With spawn_async we send the argument list exactly as it should
be run, which is safe against problems with encoding and escape
characters.

We also use spawn_async_notify_as when we launch programs that follow
the startup-notify protocol.