thunar: Implement CopyTo, MoveTo, Update for Thunar 1.2
Implement Thunar actions CopyTo, MoveTo that are simple & nice
commands for the user to copy and move files. When these actions are
not available we show a runtime error "Thunar does not support this
operation".

The Empty Trash, Show Properties, Show in File Manager actions are all
special-cased to try both the two-arg version for 1.2 as well as the
one-argument version for Thunar 1.0