commands: Use subshells for all execution
The shell sh will parse the commandline for us and it allows the user
to use real shell pipelines. It also rids of all the pesky parsing and
escape issues.

Applies to Run (Get Output), Filter through Command, Pass to Command,
Write to Command..

For FileLeaves we still just run them straighforwardly.