ui: Workaround conflict between repeated key swallowing and ibus
Quote: """
Some more debugging and it turns out that for whatever reason, ibus
generates an extra keypress event for each letter pressed. The second
event is caught by the "held key" detection feature of Kupfer (for
launching default actions) and is getting swallowed.

...

Since the time between the two events appears to be on the order of 1
to 5 ms, attached is a patch that puts a minimum threshold on the
repeat time.
I have it set to 20ms, which is very safe for making sure ibus
keypresses get through, although it will break things for people who
have their repeat rate set to 50 Hz.
The ideal patch would compare the delay to the current X11 repeat rate
setting, but this gets things working.
"""

Launchpad-bug: https://bugs.launchpad.net/kupfer/+bug/601816