Support Relaying Keyboard events from a different $DISPLAY (!!)
libkeybinder at the moment can only work with one root window "the
default display and screen for the process".

To circum-hack this we provide a small little tool "kupferrelay.py"
that relays global keyboard events (only the ones that are configured
in Kupfer) to the main instance. All the kupfer instances have to be
on separate $DISPLAYS but the same Dbus session bus.

Note that not all actions in Kupfer support launching on the correct
display. In fact, "Launch Application" is almost the only one working.