clipfiles: Implement Clipboard File proxy object (And Clipboard Text)
Provide a proxy object for the files that are (possibly) on the current
clipboard.

We want integration with file managers everywhere, and they all support
copying files to the clipboard--this method lets you focus them easily
in kupfer.

We also implement the current clipboard text as a proxy object as the
same time.