Remove art/icon-list and use only plugin-installed icons
Also, plugin-installed icons default to being skipped if they conflict
with an already themed icon (from the normal theme). The plugin can add
the special flag ``!override`` to each line in the icon-list to request
the icon always to be installed.

Moving object.svg and objects.svg into the ``core`` plugin, and now we
only install the kupfer-object icon when it's not provided by the icon
theme.