a Custom Session Management plugin
Adds a text entry for each command in the plugin preferences (not very
aligned, though - is there a way to fix that?).

The Gnome/Xfce plugins launch a dialogue for log out/shut down with
with further options, so this commit renames the classes in
session_support for these and creates extra classes for each
individual command they each cover. (That is, log out covers log out and
switch user, and shut down covers shut down, reboot and suspend.)