update prefs plugins sidebar on keyboard events
The sidebar on the preferences plugins tab (containing plugin
information) only updates on mouse actions, but should also update
to reflect activation/deactivation of plugins by the keyboard
(space, enter) - and maybe other means I don't know of.

The way I've done this means the sidebar update function gets called
twice if you click a checkbox with the mouse - is there an easy way
around that? Does it matter?