ENH: Use a much stricter set of builtins for plugins' fake load
When loading a plugin just for the info (“fake load”), use a set of
builtins that only exports the gettext _() function and nothing more.

This is not a security measure, it just catches bad plugins or plugins
that don't define the info variables at the top (or if there are random
python scripts in the plugin directories).