ENH: Fix how star and arrow columns line up in the result list
Split the info column into favourite and arrow columns, so that
favourite stars and content arrows line up nicely

Mark the main column (item name and description) to expand, so that it
uses all the rest space, and pushes the star and arrow column all the
way to the right.

This means we don't have to specify a width for the name and description
field, and it expands to the remaining space.