API: Switch Kupfer to use a new D-Bus name
Old service name was se.kaizer.kupfer which is defunct. New name uses
a domain name that the project controls.

New is as follows:

SERVICE="io.github.kupferlauncher"
OBJ="/io/github/kupferlauncher"
IFACE="io.github.kupferlauncher.Listener"

Kupfer will continue to listen to the old D-Bus service name as well
**but only for a short while**. It is a time of rapid change and now is
a good time to do the transition.

The methods that stay on have the same signatures, the methods that were
unused have not been moved to the new interface.

Fixes #38