ENH: Allow running Kupfer without Keybinder
Keybinder is the only way to configure global keybindings inside kupfer
itself, but it is old technology. In Wayland and futuristic
environments, it's not available.

If you don't use Keybinder, simply configure your window manager or
equivalent to add a global keybinding that starts the `kupfer`
executable.

In fact, Keybinder will crash if used under Wayland. The users that have
it installed anyway must run Kupfer with KUPFER_NO_KEYBINDER=1 set in
the environment.